backend/pyproject.toml
[tool.black]
line-length = 88
exclude = 'migrations'



[tool.isort]
profile = "black"

[tool.poetry]
name = "bakery"
version = "0.1.0"
description = ""
authors = ["Rajinder Mohan <director@avioxtechnologies.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.8.1"
black = "24.8.0"
certifi = "2024.8.30"
cfgv = "3.4.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
distlib = "0.3.8"
django = "5.1.1"
django-stubs = "5.0.4"
django-stubs-ext = "5.0.4"
djangorestframework = "3.15.2"
djangorestframework-stubs = "3.15.0"
filelock = "3.15.4"
flake8 = "7.1.1"
identify = "2.6.0"
idna = "3.8"
isort = "5.13.2"
mccabe = "0.7.0"
mypy = "1.11.2"
mypy-extensions = "1.0.0"
nodeenv = "1.9.1"
packaging = "24.1"
pathspec = "0.12.1"
platformdirs = "4.2.2"
pre-commit = "3.8.0"
pre-commit-hooks = "4.6.0"
pycodestyle = "2.12.1"
pyflakes = "3.2.0"
python-dotenv = "1.0.1"
pyyaml = "6.0.2"
requests = "2.32.3"
ruamel-yaml = "0.18.6"
ruamel-yaml-clib = "0.2.8"
sqlparse = "0.5.1"
types-pyyaml = "6.0.12.20240808"
types-requests = "2.32.0.20240905"
typing-extensions = "4.12.2"
urllib3 = "2.2.2"
virtualenv = "20.26.3"
twilio = "^9.3.0"
rest-framework-simplejwt = "^0.0.2"
pillow = "^10.4.0"
types-pillow = "^10.2.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
barcode = "^1.0.4"
email-validator = "^2.2.0"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"
psycopg = "^3.2.3"
django-cors-headers = "^4.5.0"
django-ckeditor = "^6.7.1"
django-filter = "^24.3"
validate-email = "^1.3"
types-pytz = "^2024.2.0.20241003"
celery = "^5.4.0"
redis = "^5.2.0"
django-celery-beat = "^2.7.0"
django-rest-swagger = "^2.2.0"
channels = "^4.1.0"
daphne = "^4.1.2"
channels-redis = "^4.2.0"
weasyprint = "^63.0"
django-case-insensitive-field = "^1.0.7"
django-multiselectfield = "^0.1.13"
xhtml2pdf = "^0.2.16"
python-barcode = "^0.15.1"
openai = "^1.63.0"
groq = "^0.18.0"
[tool.poetry.group.dev.dependencies]
types-pillow = "^10.2.0.20240822"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
